Changed how Permissions work and are assigned
Permissions are now assigned through a PermissionManager.AccountType Method with Enums having values based on their index in the enumerator.

Added test to check for this, will add more once Emanuel adds the DB Abstraction part.

PermissionManager has a method for converting int into An account type, starts at 0 goes to 3 atm, with this is a constructor that does not take an AccountType.